Restructured css files into its corresponding component folder
