+++ frontend/package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "bootstrap": "^4.5.3",
+    "hamburger-react": "^2.4.0",
     "jquery": "3.5.1",
     "js-cookie": "^2.2.1",
     "moment": "2.29.1",
