+++ frontend/src/App.js
@@ -8,14 +8,14 @@ import {
 import withContext from './Context'
 
 // Components Import
-import Header from './components/Header';
+import Header from './components/Header/Header';
 import Home from './components/Home/Home';
-import Events from './components/Events'
+import Events from './components/Events/Events'
 import About from './components/About';
 import SignUp from './components/SignUp/SignUp'
 import SignIn from './components/SignIn/SignIn'
 import NotFound from './components/NotFound'
-import Footer from './components/Footer';
+import Footer from './components/Footer/Footer';
 
 // Consumer Components
 const SignUpWithContext = withContext(SignUp)
