+++ frontend/src/components/Events/Events.js
@@ -1,4 +1,7 @@
 import React from 'react';
+
+import './styles.css'
+
 import { Calendar, momentLocalizer } from 'react-big-calendar';
 import moment from 'moment';
 import 'react-big-calendar/lib/css/react-big-calendar.css';
