+++ frontend/src/components/Events/styles.css
@@ -1,0 +1,31 @@
+/* Events */
+#events {
+   display: block;
+}
+
+   #events #current-events {
+      padding-bottom: 4em;
+      padding-top: 4em;
+   }
+
+      #events #current-events  h2 {
+         position: relative;
+         text-align: center;
+      }
+
+      #events #current-events .rbc-calendar {
+         margin-left: 2vw;
+         margin-right: 2vw;
+         box-sizing: border-box;
+         height: 95vh;
+         display: flex;
+         flex-direction: column;
+         align-items: stretch;
+      }
+
+   #events #past-events {
+      text-align: center;
+      padding-top: 2em;
+      padding-bottom: 2em;
+   }
+
