+++ frontend/src/components/Footer/Footer.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import './styles.css'
 
 export default () => (
   <footer id="footer">
