+++ frontend/src/components/Footer/styles.css
@@ -1,0 +1,93 @@
+/* Footer */
+
+#footer {
+   padding: 3em 0 3em 0;
+   opacity: 1;
+   /* background: #f6f6f6; */
+   background: #1f8ad0;
+   color: #ffffff;
+   text-align: center;
+}
+
+   #footer a {
+      -moz-transition: color 0.2s ease-in-out;
+      -webkit-transition: color 0.2s ease-in-out;
+      -ms-transition: color 0.2s ease-in-out;
+      transition: color 0.2s ease-in-out;
+      color: #ffffff;
+      text-decoration: none;
+   }
+
+      #footer a:active, #footer a:hover {
+         color: #d6d6d6;
+      }
+
+   #footer .icons {
+      font-size: 1em;
+   }
+
+      #footer .icons li {
+         padding-right: 2.5em;
+      }
+
+         #footer .icons li:last-child {
+            padding-right: 0;
+         }
+
+   #footer .footlinks {
+      margin: 2em 0;
+      padding: 0;
+      text-align: center;
+   }
+
+      #footer .footlinks li {
+         border-left: solid 1px rgba(144, 144, 144, 0.25);
+         display: inline-block;
+         list-style: none;
+         margin-left: 1.5em;
+         padding-left: 1.5em;
+         text-decoration: underline;
+      }
+
+         #footer .footlinks li:first-child {
+            border-left: 0;
+            margin-left: 0;
+            padding-left: 0;
+         }
+
+   @media screen and (max-width: 980px) {
+
+      #footer {
+         padding: 6em 0 4em 0;
+      }
+
+   }
+
+   @media screen and (max-width: 736px) {
+
+      #footer {
+         padding: 3em 0 2em 0;
+      }
+
+         #footer .icons {
+            font-size: 1em;
+         }
+
+   }
+
+   @media screen and (max-width: 480px) {
+
+      #footer {
+         padding: 3em 0 1em 0;
+      }
+
+         #footer .footlinks li {
+            border-left: 0;
+            display: block;
+            margin: 0;
+            padding: 0;
+            text-decoration: underline;
+         }
+
+   }
+
