+++ frontend/src/components/Header/styles.css
@@ -1,0 +1,203 @@
+/* Header */
+
+#header {
+   background-color: #fff;
+   border-bottom: solid 1px rgba(144, 144, 144, 0.25);
+   box-shadow: 0px 0.0375em 0.125em 0px rgba(0, 0, 0, 0.05);
+   color: #fff;
+   cursor: default;
+   font-size: 1.25em;
+   height: 4.5em;
+   left: 0;
+   line-height: 4.4em;
+   position: fixed;
+   text-transform: uppercase;
+   top: 0;
+   width: 100%;
+   z-index: 10000;
+   z-index: 10001;
+}
+
+   #header h1 {
+      color: #1f8ad0;
+      font-weight: 400;
+      height: inherit;
+      left: 1.25em;
+      line-height: inherit;
+      margin: 0;
+      padding: 0;
+      position: absolute;
+      top: 0;
+   }
+
+   #header nav {
+      height: inherit;
+      line-height: inherit;
+      position: absolute;
+      right: 1.25em;
+      top: 0;
+      vertical-align: middle;
+   }
+
+      #header nav ul {
+         list-style: none;
+         margin: 0;
+         padding-left: 0;
+      }
+
+         #header nav ul li {
+            border-radius: 4px;
+            display: inline-block;
+            padding-left: 0;
+         }
+
+            #header nav ul li a {
+               -moz-transition: color 0.2s ease-in-out;
+               -webkit-transition: color 0.2s ease-in-out;
+               -ms-transition: color 0.2s ease-in-out;
+               transition: color 0.2s ease-in-out;
+               color: #000000;
+               display: inline-block;
+               text-decoration: none;
+            }
+
+               #header nav ul li a:hover {
+                  color: #484848;
+               }
+
+            #header nav ul li:first-child {
+               margin-left: 0;
+            }
+
+            #header nav ul li .button {
+               border-color: rgba(144, 144, 144, 0.25);
+               box-shadow: none;
+               height: 3em;
+               line-height: 2.9em;
+               margin-bottom: 0;
+               padding: 0 1.5em;
+               position: relative;
+               top: -0.075em;
+               vertical-align: middle;
+            }
+
+         #header #mobile-ul #SignIn a {
+            background-color: #1f8ad0;
+            position: relative;
+            color: #ffffff;
+            width: 5em;
+            bottom: -100%;
+         }
+
+   #header .container {
+      position: relative;
+   }
+
+      #header .container h1 {
+         left: 0;
+      }
+
+      #header .container nav {
+         right: 0;
+      }
+
+   #header.alt {
+      background-color: #ffffff;
+      border: 0;
+      width: 100vw;
+      box-shadow: none;
+      height: 7em;
+      line-height: 3.5em;
+      position: absolute;
+      box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
+   }
+
+      #header.alt #nav-mobile {
+         display: none;
+      }
+
+      #header.alt #nav1 {
+         top: 1.75em;
+         left: 2em;
+         max-width: 15em;
+         text-align: left;
+      }
+
+         #header.alt #nav1 a {
+            color: #1f8ad0;
+            width: 7em;
+            text-align: left;
+            display: block;
+         }
+
+         #header.alt #nav1 a:active, #header.alt #nav1 a:hover {
+            color: #6eb2df;
+         }
+      
+      #header.alt #nav2 {
+         top: 1.75em;
+         right: 2em;
+      }
+         #header.alt #nav2 a {
+            background-color: #1f8ad0;
+            color: #ffffff;
+            width: 5em;
+            text-align: center;
+            display: block;
+         }
+
+         #header.alt #nav2 a:active, #header.alt #nav2 a:hover {
+            background-color: #6eb2df;
+            color: #ffffff;
+         }
+
+      #header.alt #logo {
+         top: 0em;
+         text-align: center;
+         left: 0em;
+      }
+
+         #header.alt #logo img {
+            height: 7em;
+         }
+
+      #header.alt h1 a {
+         color: #1f8ad0;
+      }
+
+      #header.alt .button {
+         border-color: rgba(255, 255, 255, 0.5);
+         color: #1f8ad0 !important;
+      }
+
+   @media screen and (max-width: 980px) {
+
+      #header.alt {
+         height: 4em;
+         z-index: 0;
+      }
+
+         #header.alt #nav1 {
+            display: none;
+         }
+
+         #header.alt #nav2 {
+            display: none;
+         }
+
+         #header.alt #logo {
+            top: 0em;
+            text-align: left;
+            left: 0em;
+         }
+
+            #header.alt #logo img {
+               height: 4em;
+            }
+         
+         #header.alt #nav-mobile {
+            display: inline;
+            top: 0.75em;
+            max-width: 3em;
+         }
+   }
