+++ frontend/src/components/Home/Banner.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 
 import styled from 'styled-components'
 
-import Header from '../Header'
+import Header from '../Header/Header'
 
 const BannerHeading = styled.h2`
    margin-top: 30px;
