+++ frontend/src/components/Home/styles.css
@@ -1,0 +1,258 @@
+/* One */
+
+#one div {
+   display: flex;
+
+}
+
+   #one div #textbox {
+      display: block;
+      background-color: #1f8ad0;
+      padding: 3em 4em 3em 4em;
+      margin: 2em 2em 2em 2em;
+      max-width: 45%;
+      position: relative;
+      top: 0em;
+   }
+
+      #one div #textbox h2 {
+         color: #ffffff;
+      }
+
+      #one div #textbox p {
+         color: #ffffff;
+      }
+
+      #one div #textbox a input {
+         color: #1f8ad0;
+         background-color: #ffffff;
+      }
+
+      #one div #textbox a input:hover {
+         color: #1f8ad0;
+         background-color: #aaa;
+      }
+
+      #one div #textbox a input:hover {
+         text-decoration: none;
+      }
+
+   #one div #images {
+      position: relative;
+      top: 0em;
+      margin-left: -5em;
+      z-index: -1;
+   }
+
+@media screen and (max-width: 980px) {
+   #one div {
+      display: block;
+      text-align: center;
+   }
+
+      #one div #textbox {
+         max-width: 100%;
+         margin: 2em 2em 2em 2em;
+      }
+
+      #one div #images {
+         display: none;
+      }
+}
+
+@media screen and (max-width: 480px) {
+   #one div #textbox {
+      max-width: 100%;
+      margin: 0.5em 0.5em 0.5em 0.5em;
+   }
+}
+
+/* Two */
+
+#two div {
+   display: flex;
+   
+}
+
+   #two div #textbox {
+      display: block;
+      background-color: #1f8ad0;
+      padding: 3em 4em 3em 4em;
+      margin: 2em 2em 2em -5em;
+      max-width: 45%;
+      position: relative;
+      top: 0em;
+   }
+
+      #two div #textbox h2 {
+         color: #ffffff;
+      }
+
+      #two div #textbox p {
+         color: #ffffff;
+      }
+
+      #two div #textbox a input {
+         background-color: #ffffff;
+         color: #1f8ad0;
+      }
+
+      #two div #textbox a input:hover {
+         color: #1f8ad0;
+         background-color: #aaa;
+         text-decoration: none;
+      }
+
+   #two div #images {
+      position: relative;
+      top: 0em;
+      margin-left: 5em;
+      z-index: -1;
+   }
+
+@media screen and (max-width: 980px) {
+   #two div {
+      display: block;
+      text-align: center;
+   }
+
+      #two div #textbox {
+         max-width: 100%;
+         margin: 2em 2em 2em 2em;
+      }
+
+      #two div #images {
+         display: none;
+      }
+}
+
+@media screen and (max-width: 480px) {
+   #two div #textbox {
+      max-width: 100%;
+      margin: 0.5em 0.5em 0.5em 0.5em;
+   }
+}
+
+/* Instagram */
+#instagram {
+   position: relative;
+   text-align: center;
+   padding-left: 15em;
+   padding-right: 15em;
+   padding-bottom: 3em;
+}
+
+   #instagram h2 {
+      position: relative;
+      text-align: center;
+   }
+
+   #instagram div {
+      display: inline;
+      position: relative;
+   }
+
+      #instagram div div a {
+         z-index: 1;
+      }
+
+      #instagram div div img {
+         position: relative;
+         margin: 5px;
+         height: 20vw;
+         width: 20vw;
+      }
+
+@media screen and (max-width: 860px) {
+   #instagram {
+      padding-left: 10em;
+      padding-right: 10em;
+   }
+
+      #instagram h2 {
+         font-size: 30px;
+      }
+}
+
+@media screen and (max-width: 560px) {
+   #instagram {
+      padding-left: 8em;
+      padding-right: 8em;
+   }
+
+      #instagram h2 {
+         font-size: 20px;
+      }
+}
+
+@media screen and (max-width: 480px) {
+   #instagram {
+      padding-left: 5em;
+      padding-right: 5em;
+   }
+}
+
+/* Banner */
+
+body.is-touch #banner {
+   background-attachment: scroll;
+}
+
+#banner {
+   position: relative;
+   height: 100vh;
+   padding-top: 50vh;
+   padding-bottom: 50vh;
+   background-image:  url("../../styles/css/images/overlay.png"), url("../../images/ibsa_cover.jpg");
+   background-position: center top;
+   background-size: cover;
+   line-height: 1.75;
+   text-align: center;
+}
+
+   #banner:last-child {
+      margin-bottom: 0;
+   }
+
+   #banner h2 {
+      color: #ffffff;
+      display: inline-block;
+      font-size: 4.0em;
+      line-height: 1.35;
+      margin-bottom: 0.0em;
+   }
+
+   #banner p {
+      color: #aaa;
+      font-size: 1.5em;
+      margin-bottom: 1.75em;
+      text-transform: uppercase;
+   }
+
+   @media screen and (max-width: 1280px) {
+
+         #banner h2 {
+            font-size: 2.5em;
+         }
+
+   }
+
+   @media screen and (max-width: 736px) {
+
+         #banner br {
+            display: none;
+         }
+
+         #banner h2 {
+            font-size: 2.25em;
+         }
+
+         #banner p {
+            font-size: 1.25em;
+         }
+
+   }
+
+   @media screen and (max-width: 480px) {
+
+   }
